refac: adiciona controllers e models para iterator
