+++ backend/api/src/main/java/com/api/API/models/Restaurante.java
@@ -1,4 +1,19 @@
 package com.api.API.models;
 
 public class Restaurante {
+    private String nome;
+    private Menu menu;
+
+    public Restaurante(String nome, int capacidadeMenu) {
+        this.nome = nome;
+        this.menu = new Menu(capacidadeMenu);
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public Menu getMenu() {
+        return menu;
+    }
 }
