--- backend/api/src/main/java/com/api/API/models/iterator/Aplicacao/Main.java
@@ -1,37 +1,0 @@
-package Aplicacao;
-
-import modelo.entidades.Prato;
-import modelo.entidades.Restaurante;
-import modelo.iterador.Iterador;
-
-public class Main {
-    public static void main(String[] args) {
-        // Criando restaurante e menu
-        Restaurante restaurante = new Restaurante("Restaurante Chef Indica", 5);
-
-        // Adicionando pratos ao menu
-        restaurante.getMenu().adicionarPrato(new Prato("Spaghetti Carbonara", 35.50));
-        restaurante.getMenu().adicionarPrato(new Prato("Risoto de Cogumelos", 42.00));
-        restaurante.getMenu().adicionarPrato(new Prato("File Mignon ao Molho Madeira", 65.90));
-        restaurante.getMenu().adicionarPrato(new Prato("Salmao Grelhado", 58.00));
-        restaurante.getMenu().adicionarPrato(new Prato("Tiramisu", 25.00));
-
-        // Obtendo iterador do menu
-        Iterador iterador = restaurante.getMenu().criarIterador();
-
-        // Percorrendo o menu com o iterador
-        System.out.println("Menu do Restaurante " + restaurante.getNome() + ":");
-        while (iterador.temProximo()) {
-            Prato prato = (Prato) iterador.proximo();
-            System.out.println("- " + prato);
-        }
-
-        // Reiniciando o iterador
-        iterador.reiniciar();
-        System.out.println("\nIteracao reiniciada:");
-        while (iterador.temProximo()) {
-            Prato prato = (Prato) iterador.proximo();
-            System.out.println("- " + prato);
-        }
-    }
-}
