+++ backend/api/src/main/java/com/api/API/models/iterator/ColecaoIterador.java
