--- backend/api/src/main/java/com/api/API/models/iterator/Entidades/Restaurante.java
@@ -1,19 +1,0 @@
-package modelo.entidades;
-
-public class Restaurante {
-    private String nome;
-    private Menu menu;
-
-    public Restaurante(String nome, int capacidadeMenu) {
-        this.nome = nome;
-        this.menu = new Menu(capacidadeMenu);
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public Menu getMenu() {
-        return menu;
-    }
-}
